Move function used to create the home page from App.js
